After testing, it was found that the fit of the leadscnut was far too constrained.
There was also significant flex when bending downwards.
To solve this, steel L brackets will be added similar to the z block.
The magnetic pin/ bed position may also be lowered to allow the overall thickning of the plastic parts that hold the arms, and intergration of the top 2 screws into it for even more stifness.
This lowering will reduce the z movement by 5-10mm, but its worth it.

There was also noticable side to side wiggle of the arms in their socket.
This will be because of no constrainment by the screwed top bit, so a wedge will be put in place to attempt to constrain their side to side motion better.

Simulation was done (https://www.simscale.com/projects/harryloud/bending_sim/) on the default positron aluminium arms and carage to see what the displacement is of current.
At 100N, the arms will displace 1.6mm.
Thus our target sould be ~2mm for this version.

Signed-off-by: Harry Loud <harryloud@yahoo.co.uk>