Z Block & 90 deg bed holder tested and stable.
Stainless bracket installed in the 90 deg bed holder for improoved stifness.
Simulation showed this reduced bending from 4mm - ~2.8mm (100N), which is an improovment on v3.
So essentialy its just the extrusion arms that bend now, the 90 degree holder is ridgid.
https://www.simscale.com/projects/harryloud/bending_again/   (V4)

This part is complete and can move to enclosure design.

Am thinking about placing the blower fan on the z extrusion, the otehr side of the extruder (leadscrew side) and use long brackets to hold it firm.
This allows us to sit the fanontop of components like tehe xtruder does, whilst freeing up room inside the enclosure for wiring, the psu etc.

Signed-off-by: Harry Loud <harryloud@yahoo.co.uk>